받지마세여

+++ src/support/model/vo/Support.java
@@ -12,7 +12,7 @@ public class Support {
 	private int needSupport;
 	private int supportHuman;
 	private String picPath;
-	private long picSize;
+	private double picSize;
 	private String userId;
 	private int percentage;
 	
@@ -110,7 +110,7 @@ public class Support {
 		return picSize;
 	}
 
-	public void setPicSize(long picSize) {
+	public void setPicSize(double picSize) {
 		this.picSize = picSize;
 	}
